+++ Test/Stub/ConektaQuoteRepositoryFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Conekta\Payments\Model;
+
+class ConektaQuoteRepositoryFactory
+{
+    public function create()
+    {
+        return null;
+    }
+}